requirements.txt
amqp==2.3.2
arabic-reshaper==2.0.14
Babel==2.6.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
configparser==3.5.0
crypto==1.4.1
cycler==0.10.0
Django==2.1.2
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-js-asset==1.1.0
django-mysql==2.4.1
django-querysetsequence==0.10
django-redis==4.9.0
django-redis-cache==1.7.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
flower==0.9.2
future==0.17.0
idna==2.7
jdatetime==3.0.2
kiwisolver==1.0.1
kombu==4.2.1
matplotlib==3.0.1
multidict==4.4.2
mysqlclient==1.3.13
Naked==0.1.31
numpy==1.15.3
persian-wordcloud==1.3.1
Pillow==5.3.0
pycryptodome==3.7.0
PyJWT==1.6.4
pyparsing==2.2.2
python-bidi==0.4.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
redis==2.10.6
requests==2.20.0
shellescape==3.4.1
six==1.11.0
sqlparse==0.2.4
tornado==5.1.1
urllib3==1.24
swagger-spec-validator==2.4.3
uWSGI==2.0.17.1
vine==1.1.4
wordcloud==1.5.0
